scripts/requirements.txt
importlib_metadata==8.7.0
importlib_resources==6.5.2
libusb-package==1.0.26.3
packaging==25.0
pkg_about==1.3.6
pyusb==1.3.1
zipp==3.23.0
